requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
Werkzeug==2.1.2
zipp==3.8.0
